Autogenerate Native protocol specifications documentation
Uses cqlprotodoc from cassandra-website to generate the versioned native protocol spec html pages.
Uses process-native-protocol-specs-in-docker.sh as a wrapper around this to also generate the native protocol spec adoc page  that includes all the versions and makes each header navigable. This is executed as part of the existing `ant gen-asciidoc` and doc/Makefile.

 patch by Olexandr Luzhniy; reviewed by Mick Semb Wever for CASSANDRA-19915